requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
python-dotenv==1.0.1
gunicorn==23.0.0
